ADD: split some funcs in views to a util func
